resources->fobtap point, stop floating point on stock, activate on resource stock

+++ src/mutations/resourcesMuts.js
@@ -8,14 +8,14 @@ function resourcesMuts(resources, ev){
 		case "resource-used":
 			resources.forEach( resource => {
 				if (resource.resourceId == ev.resourceId){
-					resource.stock -= parseFloat(ev.amount)
+					resource.stock -= parseInt(ev.amount)
 				}
 			})
 			break
 		case "resource-stocked":
 			resources.forEach( resource => {
 				if (resource.resourceId == ev.resourceId){
-						resource.stock += parseFloat(ev.amount)
+						resource.stock += parseInt(ev.amount)
 				}
 			})
 			break
